Remove internal line breaks in picks.csv
I highlighted this issue on github and in stack exchange (https://stackoverflow.com/questions/62413052/unwanted-newlines-for-a-file-write-csv-when-entries-are-overfull-too-long).

When the probability threshold was set low such that PhaseNet made many picks, some of the column entries would become overfull and it would automatically insert a line break and bump subsequent entries to a new row. This led to all sort of complications later, especially when reading the csv with pandas. The stack exchange post clarifies the problem and the solution which I have added in this pull request.