Cleaned picks.csv formatting with the csv library
I described this issue in under https://github.com/wayneweiqiang/PhaseNet/issues/9 and here is the PR with the fix. The fix is a two-part solution. 

First, I open the picks.csv file as fclog with the csv library and write the header row. I then open picks.csv in the append mode and write the results to picks.csv batch by batch. The results are converted from arrays to lists (this removes the empty white spaces), and then written to the row in a list of results instead of a single string formatted with the results. This also fixes the new-line issue I had opened and resolved earlier. 

I have tried and tested the method and it works as hoped. 

Best,
Lenni